DEV: Change unit test to acceptance test
